test: 관련된 테스트 코드를 각 파일로 분리
- lottoCreation.test.js: 로또 생성과 관련된 테스트
- winningNumbers.test.js: 지난주 당첨 번호와 관련된 테스트
- lottoResult.test.js: 당첨 결과와 관련된 테스트
